REPORTER_WORKER/requirements.txt
PyJWT~=2.9.0
Flask~=3.0.3
Werkzeug~=3.0.4
psycopg2~=2.9.9

pandas
matplotlib
numpy